refactor: folder layout and cva import
