+++ src/components/ui/Alert/Alert.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { AlertIcon } from './alertIcon';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../../utils/cn';
 
 type AlertType = 'success' | 'error' | 'info' | 'warning' | 'default';
