+++ src/components/ui/Avatar/Avatar.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { cn } from '../../../lib/utils';
+import { cn } from '../../../utils/cn';
 import useMediaQuery from '../../../hooks/useMediaQuery';
 
 export interface AvatarProps {
