+++ src/components/ui/accordion/components/AccordionItem.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { cn } from '../../../../lib/utils';
+import { cn } from '../../../../utils/cn';
 import { accordionStyles } from '../helper/styles';
 import { AccordionItemInternalProps } from '../types';
