+++ src/styles/index.css
@@ -1,4 +1,4 @@
-@import url(./assets/font.css);
+@import url(./font.css);
 @import url('https://fonts.googleapis.com/css2?family=Figtree:ital,wght@0,300..900;1,300..900&display=swap');
 @tailwind base;
 @tailwind components;
