+++ src/utils/helper/slugify.helper.ts
@@ -1,5 +1,13 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-export function slugify(text: any) {
+/**
+ * Converts a given string into a URL-friendly slug.
+ * @param {string} text - The input string to be slugified.
+ * @returns {string} The slugified version of the input string.
+ *
+ * @example
+ * slugify("This is a test.");
+ * Returns "this-is-a-test"
+ */
+export function slugify(text: string): string {
   return text
     .toString()
     .toLowerCase()
